cpack: Build a zip for Windows
windeployqt bundles up everything needed to run acrostica on another
computer.  However, in order to get it included into the package that
cpack builds, install() needs to be used.

Make windeployqt dump the files into an intermediate directory, which
then allows use of the generic "install(DIRECTORY ...)" form to move the
files into the same directory as acrostica.exe.