Convert MissingLetters from a Table to a List
The table{model,view} was being used solely to get the structured
display.  However, that is inflexible when the view size needs to change
and is the wrong reasoning for using a table.

Switching to a list simplifies the code and also resolves one of the
main issues with non-ASCII characters by allowing the Model size to
dynamically change.  Previously, it was fixed to be proximal to the
size of the ASCII letters.